refactor(creds): extract _blind_cache_remedy to its own module
_pick_healthy.py sat at 518 lines against the 512-line cap. The remedy
builder is a pure message function -- it takes a diagnosis and returns
prose, sharing no state with the ranking logic -- so it moves out cleanly
and leaves _pick_healthy.py at 429.

Behaviour-preserving: _pick_healthy re-imports the name, so the call site
and every existing test path are unchanged. 152 tests pass across
tests/scitex_agent_container/_creds/ and _account/test_quota_cache.py.

Deliberately NOT included: a "foreign-schema" cache state I had drafted on
this branch to flag email-keyed quota-cache entries as a second writer's.
The premise is wrong. tests/.../_account/test_quota_cache.py pins an
email-keyed SAMPLE as "the lead's confirmed wire shape", and
read_quota_entry does not key on the dict key at all -- it iterates
accounts.values() and matches the `short` field. Email keys are therefore
both documented and harmless to the reader, so the detector would have
flagged the contract as a defect. The blind-pick cause named on
quota-cache-two-writers-schema-collision-blocks-all-restarts-20260802 is
retracted there with this reasoning; the incident is unexplained again.
